+++ ere-guests/Cargo.toml
@@ -122,10 +122,11 @@ zero_sized_map_values = "warn"
 
 [workspace.dependencies]
 guest-libs = { path = "libs" }
-reth-guest = { path = "stateless-validator/reth/guest" }
+ere-reth-guest = { path = "stateless-validator/reth/guest" }
 
 sp1-zkvm = "5.0.5"
 pico-sdk = { git = "https://github.com/brevis-network/pico.git", tag = "v1.1.7" }
+ziskos = { git = "https://github.com/0xPolygonHermez/zisk.git" }
 
 # branch is kw/zkevm-benchmark-workload-repo
 # NOTE: We are using a branch of a branch that has not yet been merged into master.
