+++ ere-guests/stateless-validator/reth/sp1/Cargo.toml
@@ -8,9 +8,7 @@ license = "MIT OR Apache-2.0"
 
 [dependencies]
 sp1-zkvm.workspace = true
-guest-libs.workspace = true
-sparsestate.workspace = true
-reth-guest.workspace = true
+ere-reth-guest.workspace = true
 
 reth-stateless.workspace = true
 reth-ethereum-primitives = { workspace = true, features = [
@@ -21,8 +19,6 @@ reth-primitives-traits = { workspace = true, features = [
     "serde",
     "serde-bincode-compat",
 ] }
-reth-evm-ethereum.workspace = true
-reth-chainspec.workspace = true
 
 revm = { workspace = true, default-features = false, features = [
     "kzg-rs",
