+++ ere-guests/stateless-validator/reth/sp1/src/main.rs
@@ -4,10 +4,44 @@
 
 extern crate alloc;
 
-use reth_guest::{guest::ethereum_guest, sdk::sp1::SP1SDK};
+use ere_reth_guest::{
+    guest::ethereum_guest,
+    sdk::{CycleScope, SDK},
+};
+use k256::ecdsa::VerifyingKey;
+use reth_stateless::StatelessInput;
 use tracing_subscriber::fmt;
 
 sp1_zkvm::entrypoint!(main);
+
+/// SDK implementation for SP1.
+#[allow(missing_debug_implementations)]
+pub struct SP1SDK;
+
+impl SDK for SP1SDK {
+    fn read_inputs() -> (StatelessInput, Vec<VerifyingKey>) {
+        let input = sp1_zkvm::io::read::<StatelessInput>();
+        let public_keys = sp1_zkvm::io::read::<Vec<VerifyingKey>>();
+        (input, public_keys)
+    }
+
+    fn commit_outputs(block_hash: [u8; 32], parent_hash: [u8; 32], is_valid: bool) {
+        sp1_zkvm::io::commit(&block_hash);
+        sp1_zkvm::io::commit(&parent_hash);
+        sp1_zkvm::io::commit(&is_valid);
+    }
+
+    fn cycle_scope(scope: CycleScope, message: &str) {
+        match scope {
+            CycleScope::Start => {
+                println!("cycle-tracker-report-start: {message}")
+            }
+            CycleScope::End => {
+                println!("cycle-tracker-report-end: {message}")
+            }
+        }
+    }
+}
 
 /// Entry point.
 pub fn main() {
