+++ ere-guests/stateless-validator/reth/zisk/src/main.rs
@@ -2,89 +2,54 @@
 
 #![no_main]
 
-use std::{error::Error, io::Cursor, sync::Arc};
+use std::io::Cursor;
 
-use alloy_primitives::FixedBytes;
+use ere_reth_guest::{
+    guest::ethereum_guest,
+    sdk::{CycleScope, SDK},
+};
 use k256::ecdsa::VerifyingKey;
-use reth_chainspec::ChainSpec;
-use reth_ethereum_primitives::Block as EthBlock;
-use reth_evm_ethereum::EthEvmConfig;
-use reth_primitives_traits::Block;
-use reth_stateless::{ExecutionWitness, Genesis, StatelessInput, stateless_validation_with_trie};
+use reth_stateless::StatelessInput;
 use sha2::{Digest, Sha256};
-use sparsestate::SparseState;
 
 ziskos::entrypoint!(main);
 
-/// Entry point.
-pub fn main() {
-    println!("start read_input");
-    let mut input_bytes = Cursor::new(ziskos::read_input());
-    let input: StatelessInput = bincode::deserialize_from(&mut input_bytes).unwrap();
-    let public_keys: Vec<VerifyingKey> = bincode::deserialize_from(&mut input_bytes).unwrap();
-    let genesis = Genesis {
-        config: input.chain_config.clone(),
-        ..Default::default()
-    };
-    let chain_spec: Arc<ChainSpec> = Arc::new(genesis.into());
-    let evm_config = EthEvmConfig::new(chain_spec.clone());
-    println!("end read_input");
+/// SDK implementation for Zisk.
+#[allow(missing_debug_implementations)]
+pub struct ZiskSDK;
 
-    println!("start public_inputs_preparation");
-    let header = input.block.header().clone();
-    let parent_hash = input.block.parent_hash;
-    println!("end public_inputs_preparation");
+impl SDK for ZiskSDK {
+    fn read_inputs() -> (StatelessInput, Vec<VerifyingKey>) {
+        let mut input_bytes = Cursor::new(ziskos::read_input());
+        let input: StatelessInput = bincode::deserialize_from(&mut input_bytes).unwrap();
+        let public_keys: Vec<VerifyingKey> = bincode::deserialize_from(&mut input_bytes).unwrap();
+        (input, public_keys)
+    }
 
-    let res = validate_block(
-        input.block,
-        input.witness,
-        chain_spec,
-        public_keys,
-        evm_config,
-    );
+    fn commit_outputs(block_hash: [u8; 32], parent_hash: [u8; 32], is_valid: bool) {
+        let public_inputs = (block_hash, parent_hash, is_valid);
+        let public_inputs_hash = Sha256::digest(bincode::serialize(&public_inputs).unwrap());
+        public_inputs_hash
+            .chunks_exact(4)
+            .enumerate()
+            .for_each(|(idx, bytes)| {
+                ziskos::set_output(idx, u32::from_le_bytes(bytes.try_into().unwrap()))
+            });
+    }
 
-    println!("start commit_public_inputs");
-    // The public inputs are:
-    // - block_hash : [u8;32]
-    // - parent_hash : [u8;32]
-    // - successful_block_validation : bool
-    let public_inputs = match res {
-        Ok(block_hash) => (block_hash.0, parent_hash.0, true),
-        Err(err) => {
-            println!("Block validation failed: {err}");
-            (header.hash_slow().0, parent_hash.0, false)
+    fn cycle_scope(scope: CycleScope, message: &str) {
+        match scope {
+            CycleScope::Start => {
+                println!("start: {message}")
+            }
+            CycleScope::End => {
+                println!("end: {message}")
+            }
         }
-    };
-    let public_inputs_hash = Sha256::digest(bincode::serialize(&public_inputs).unwrap());
-    public_inputs_hash
-        .chunks_exact(4)
-        .enumerate()
-        .for_each(|(idx, bytes)| {
-            ziskos::set_output(idx, u32::from_le_bytes(bytes.try_into().unwrap()))
-        });
-    println!("end commit_public_inputs");
+    }
 }
 
-fn validate_block(
-    block: EthBlock,
-    witness: ExecutionWitness,
-    chain_spec: Arc<ChainSpec>,
-    public_keys: Vec<VerifyingKey>,
-    evm_config: EthEvmConfig,
-) -> Result<FixedBytes<32>, Box<dyn Error>> {
-    println!("start public_keys_validation");
-    let recovered_block =
-        guest_libs::senders::recover_block_with_public_keys(block, public_keys, &chain_spec)?;
-    println!("end public_keys_validation");
-
-    println!("start validation");
-    let block_hash = stateless_validation_with_trie::<SparseState, _, _>(
-        recovered_block,
-        witness,
-        chain_spec,
-        evm_config,
-    )?;
-    println!("end validation");
-
-    Ok(block_hash)
+/// Entry point.
+pub fn main() {
+    ethereum_guest::<ZiskSDK>();
 }
